hurrdurr

+++ src/main/java/Interfaces/IBank.java
@@ -3,7 +3,9 @@ package Interfaces;
 import models.Customer;
 import models.Merchant;
 
+import java.math.BigDecimal;
+
 public interface IBank {
 
-    void transferMoney(Customer customer, Merchant merchant, double amount);
+    void transferMoney(Customer customer, Merchant merchant, BigDecimal amount);
 }
