+++ src/main/java/models/Transaction.java
@@ -1,5 +1,6 @@
 package models;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class Transaction {
@@ -7,10 +8,10 @@ public class Transaction {
     private Date transactionDate;
     private Account creditor; // To
     private Account debtor; // From
-    private double amount;
+    private BigDecimal amount;
     private Token token;
 
-    public Transaction(Account creditor, Account debtor, double amount, Token token, Date transactionDate){
+    public Transaction(Account creditor, Account debtor, BigDecimal amount, Token token, Date transactionDate){
         this.creditor = creditor;
         this.debtor = debtor;
         this.amount = amount;
@@ -18,7 +19,7 @@ public class Transaction {
         this.transactionDate = transactionDate;
     }
 
-    public Transaction(Account creditor, Account debtor, double amount, Token token){
+    public Transaction(Account creditor, Account debtor, BigDecimal amount, Token token){
         Date transactionDate = new Date();
         this.creditor = creditor;
         this.debtor = debtor;
@@ -44,7 +45,7 @@ public class Transaction {
     }
 
 
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
